requirements.txt
openai
airtable-python.wrapper
streamlit==1.31.0
git+https://github.com/openai/whisper.git
pyaudio
code_editor